finally got rid of tree model

+++ AUI.Views/src/AUI/Util/AMetric.h
@@ -98,13 +98,13 @@ public:
 
 inline AMetric operator"" _px(unsigned long long v)
 {
-	return AMetric(static_cast<float>(v), AMetric::T_PX);
+	return AMetric(static_cast<float>(static_cast<long long>(v)), AMetric::T_PX);
 }
 inline AMetric operator"" _dp(unsigned long long v)
 {
-	return AMetric(static_cast<float>(v), AMetric::T_DP);
+	return AMetric(static_cast<float>(static_cast<long long>(v)), AMetric::T_DP);
 }
 inline AMetric operator"" _pt(unsigned long long v)
 {
-	return AMetric(static_cast<float>(v), AMetric::T_PT);
+	return AMetric(static_cast<float>(static_cast<long long>(v)), AMetric::T_PT);
 }
